backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
chromadb==0.5.23
sentence-transformers==3.3.1
google-generativeai==0.8.3
websockets==13.1
jinja2==3.1.4
aiofiles==24.1.0
